feat(event/owneraction,event-visitors,containers/containerownersaction): refactoring OwnersAction
